+++ src/main.js
@@ -2,6 +2,7 @@ JSB.require("vendor/mnutils");
 JSB.require("vendor/mnnote");
 JSB.require("WebDevServerConfig");
 JSB.require("CommentData");
+JSB.require("UndoGroupingHelper");
 JSB.require("CommentMutations");
 JSB.require("BatchCommentActions");
 JSB.require("WebBridgeCommands");